+++ data-browser/src/components/forms/ResourceSelector.tsx
@@ -160,12 +160,13 @@ export const ResourceSelector = React.memo(function ResourceSelector({
 });
 
 /** For a given class URL, this tries to return a URL of a Collection containing these. */
+// TODO: Scope to current store / make adjustable https://github.com/atomicdata-dev/atomic-data-browser/issues/295
 export function getCollectionURL(classtypeUrl?: string): string | undefined {
   switch (classtypeUrl) {
     case urls.classes.property:
-      return urls.properties.getAll;
+      return 'https://atomicdata.dev/properties/?page_size=999';
     case urls.classes.class:
-      return urls.classes.getAll;
+      return 'https://atomicdata.dev/classes/?page_size=999';
     case urls.classes.agent:
       return 'https://atomicdata.dev/agents/';
     case urls.classes.commit: