+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This changelog covers all three packages, as they are (for now) updated as a who
 - Add `Store.postToServer` method, add `endpoints`, `importJsonAdString`
 - Add new file preview UI for images, audio, text and PDF files.
 - Add new file preview types to the folder grid view.
+- Add `store.preloadClassesAndProperties` and remove `urls.properties.getAll` and `urls.classes.getAll`. This enables using `atomic-data-browser` without relying on `atomicdata.dev` being available.
 
 ## v0.35.0